Fix some bugs in virtualized compiler.
These bugs are related to the refactoring of context errors, and have
been present since early M2. The virtualized compiler relied on some
silent failures to not set bad type information because the errors
were caught and cleaned at the top-level of typed(), but this is not
the case anymore. The fix is to explicitly set context.retyping=true,
allowing the virtualized compiler to change bad type information.

These bugs were fixed in 2.10.0-RC1, but sadly crept back afterwards,
because the fix was not merged back into the mainline topic-virt.